% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABR.R
\name{CalcNormalImpvolSabrHagan}
\alias{CalcNormalImpvolSabrHagan}
\title{Hagan's formula calculating normal model implied vol under SABR model. Currently support only beta=0}
\usage{
CalcNormalImpvolSabrHagan(forward, strike = forward, t.exp = 1, sigma0,
  alpha = 0, rho = 0)
}
\arguments{
\item{forward}{forward stock price}

\item{strike}{strike price}

\item{t.exp}{time to expiry}

\item{sigma0}{sigma0}

\item{alpha}{alpha vol of vol}

\item{rho}{rho correlation}
}
\value{
normal implied vol
}
\description{
Hagan's formula calculating normal model implied vol under SABR model. Currently support only beta=0
}
\examples{
sigma0 <- 0.68/100
alpha <- 0.3691
rho <- -0.0286
forward <- 4.35/100
strike <- c( 4, seq(4.05,4.95,0.1), 5)/100
t.exp <- 10
vol <- phbsasp::CalcNormalImpvolSabrHagan(strike=strike, forward=forward, sigma0=sigma0, t.exp=t.exp, alpha=alpha, rho=rho)
price <- phbsasp::CalcNormalPrice(forward=forward, strike=strike, t.exp=t.exp, sigma=vol)
}

